flask/requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
deep-translator==1.11.4
Deprecated==1.2.14
Flask==3.0.0
google-ai-generativelanguage==0.1.0
google-api-core==2.14.0
google-auth==2.23.4
google-generativeai==0.1.0rc1
googleapis-common-protos==1.61.0
googletrans==4.0.0rc1
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
redis==4.1.1
requests==2.31.0
rfc3986==1.5.0
rq==1.10.1
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
